# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "My API",
    "description": "Documentation automatically generated by the <b>swagger-autogen</b> module."
  },
  "host": "localhost:3000",
  "basePath": "/",
  "tags": [
    {
      "name": "Token",
      "description": "Token APIs"
    },
    {
      "name": "Image",
      "description": "Image APIs"
    }
  ],
  "schemes": [
    "http",
    "https"
  ],
  "securityDefinitions": {
    "apiKeyAuth": {
      "type": "apiKey",
      "in": "header",
      "name": "X-API-KEY",
      "description": "any description..."
    }
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/Image/Upload": {
      "post": {
        "tags": [
          "Image"
        ],
        "description": "Upload image to IPFS",
        "consumes": [
          "multipart/form-data"
        ],
        "parameters": [
          {
            "name": "image",
            "in": "formData",
            "type": "file",
            "required": true,
            "description": "Some description..."
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/Token/Mint": {
      "post": {
        "tags": [
          "Token"
        ],
        "description": "Mint NFT Token",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "description": "Token Information",
            "schema": {
              "$ref": "#/definitions/Mint"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/Token/Get/{id}": {
      "get": {
        "tags": [
          "Token"
        ],
        "description": "Mint NFT Token",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/Token/Transfer": {
      "post": {
        "tags": [
          "Token"
        ],
        "description": "Transfer NFT Token",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "description": "Transfer Information",
            "schema": {
              "$ref": "#/definitions/Transfer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    }
  },
  "definitions": {
    "Mint": {
      "type": "object",
      "properties": {
        "token_id": {
          "type": "string",
          "example": "token_nft"
        },
        "owner": {
          "type": "string",
          "example": "aura1mmsenr5nejugxqv6jtetlckwjhsmft7f8dqud4"
        },
        "token_uri": {
          "type": "string",
          "example": "ANFT"
        },
        "extension": {
          "type": "object",
          "properties": {}
        }
      },
      "required": [
        "token_id",
        "owner",
        "token_uri",
        "extension"
      ]
    },
    "Transfer": {
      "type": "object",
      "properties": {
        "token_id": {
          "type": "string",
          "example": "token_nft"
        },
        "recipient": {
          "type": "string",
          "example": "aura1n9dk7hvqtqrpxuc52w9qsmd95dm83ccjum2636"
        }
      },
      "required": [
        "token_id",
        "recipient"
      ]
    }
  }
}